Refactor: split monolithic Axis into Axis_ / Axis_X / Axis_Y
Axis.java handled both X and Y axes via an internal Direction enum and
repeated if/else branching throughout every major method. This makes
adding Y-axis-only features (e.g. merged/grouped Y-axes from PR #917)
pollute a shared class with state that has no meaning for the X-axis.

Following the established Plot_ / Legend_ pattern:
  Axis_<ST,S>  – abstract base: shared fields (min/max, dataType,
                 axisTickCalculator, bounds, axisTitle, axisTick),
                 shared methods (addMinMax, getScreenValue contract,
                 getChartValue contract), and the Direction enum.
  Axis_X<ST,S> – X-axis: preparePaint, paint, getXAxisHeightHint,
                 getAxisTickCalculatorForX, getScreenValue, getChartValue.
  Axis_Y<ST,S> – Y-axis: preparePaint, paint, getYAxisWidthHint,
                 getAxisTickCalculatorForY, getScreenValue, getChartValue.
                 Future Y-only features (merged axes, etc.) live here.

All Direction-enum references updated from Axis.Direction -> Axis_.Direction
across AxisTick, AxisTickLabels, AxisTickMarks, AxisTitle, all
AxisTickCalculator_* subclasses, Formatter_Number, Formatter_LogNumber,
PlotContent_XY, AxisPair, Chart, and test files.

All 26 existing tests pass. No behaviour change.
