+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisPair.java
@@ -20,13 +20,13 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
 
   private final Chart<ST, S> chart;
 
-  private final Axis<ST, S> xAxis;
-  private final Axis<ST, S> yAxis;
-  private final TreeMap<Integer, Axis<ST, S>> yAxisMap;
+  private final Axis_X<ST, S> xAxis;
+  private final Axis_Y<ST, S> yAxis;
+  private final TreeMap<Integer, Axis_Y<ST, S>> yAxisMap;
   private final Rectangle2D.Double leftYAxisBounds;
   private final Rectangle2D.Double rightYAxisBounds;
-  private Axis<ST, S> leftMainYAxis;
-  private Axis<ST, S> rightMainYAxis;
+  private Axis_Y<ST, S> leftMainYAxis;
+  private Axis_Y<ST, S> rightMainYAxis;
 
   /**
    * Constructor
@@ -38,9 +38,9 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     this.chart = chart;
 
     // add axes
-    xAxis = new Axis<ST, S>(chart, Axis.Direction.X, 0);
-    yAxis = new Axis<ST, S>(chart, Axis.Direction.Y, 0);
-    yAxisMap = new TreeMap<Integer, Axis<ST, S>>();
+    xAxis = new Axis_X<>(chart);
+    yAxis = new Axis_Y<>(chart, 0);
+    yAxisMap = new TreeMap<>();
     yAxisMap.put(0, yAxis);
     leftYAxisBounds = new Rectangle2D.Double();
     rightYAxisBounds = new Rectangle2D.Double();
@@ -69,8 +69,8 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     // calculate width first
     if (desiredLeftYAxisWidth > 0) {
       double widthEstimation = 0;
-      for (Entry<Integer, Axis<ST, S>> e : yAxisMap.entrySet()) {
-        Axis<ST, S> ya = e.getValue();
+for (Entry<Integer, Axis_Y<ST, S>> e : yAxisMap.entrySet()) {
+      Axis_Y<ST, S> ya = e.getValue();
         if (styler.getYAxisGroupPosistion(e.getKey()) == YAxisPosition.Right) {
           continue;
         }
@@ -95,8 +95,8 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     }
     double leftStartFirst = leftStart;
 
-    for (Entry<Integer, Axis<ST, S>> e : yAxisMap.entrySet()) {
-      Axis<ST, S> ya = e.getValue();
+    for (Entry<Integer, Axis_Y<ST, S>> e : yAxisMap.entrySet()) {
+      Axis_Y<ST, S> ya = e.getValue();
       if (styler.getYAxisGroupPosistion(e.getKey()) == YAxisPosition.Right) {
         continue;
       }
@@ -148,8 +148,8 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     int rightCount = 0;
 
     // traverse reverse
-    for (Entry<Integer, Axis<ST, S>> e : yAxisMap.descendingMap().entrySet()) {
-      Axis<ST, S> ya = e.getValue();
+    for (Entry<Integer, Axis_Y<ST, S>> e : yAxisMap.descendingMap().entrySet()) {
+      Axis_Y<ST, S> ya = e.getValue();
       if (styler.getYAxisGroupPosistion(e.getKey()) != YAxisPosition.Right) {
         continue;
       }
@@ -249,13 +249,13 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
         if (yAxisMap.containsKey(yIndex)) {
           continue;
         }
-        yAxisMap.put(yIndex, new Axis<ST, S>(chart, Axis.Direction.Y, yIndex));
+        yAxisMap.put(yIndex, new Axis_Y<>(chart, yIndex));
       }
     }
 
     // set the axis data types, making sure all are compatible
     xAxis.setDataType(null);
-    for (Axis<ST, S> ya : yAxisMap.values()) {
+    for (Axis_Y<ST, S> ya : yAxisMap.values()) {
       ya.setDataType(null);
     }
     for (S series : chart.getSeriesMap().values()) {
@@ -279,7 +279,7 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
 
     // calculate axis min and max
     xAxis.resetMinMax();
-    for (Axis<ST, S> ya : yAxisMap.values()) {
+    for (Axis_Y<ST, S> ya : yAxisMap.values()) {
       ya.resetMinMax();
     }
 
@@ -313,7 +313,7 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     }
 
     overrideMinMaxForXAxis();
-    for (Axis<ST, S> ya : yAxisMap.values()) {
+    for (Axis_Y<ST, S> ya : yAxisMap.values()) {
       overrideMinMaxForYAxis(ya);
     }
 
@@ -323,7 +323,7 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
           "Series data (accounting for error bars too) cannot be less or equal to zero for a logarithmic X-Axis!!!");
     }
     if (chart.getStyler().isYAxisLogarithmic()) {
-      for (Axis<ST, S> ya : yAxisMap.values()) {
+      for (Axis_Y<ST, S> ya : yAxisMap.values()) {
         if (ya.getMin() <= 0.0) {
           // System.out.println(getMin());
           throw new IllegalArgumentException(
@@ -336,7 +336,7 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
       throw new IllegalArgumentException(
           "Series data (accounting for error bars too) cannot be equal to Double.POSITIVE_INFINITY!!!");
     }
-    for (Axis<ST, S> ya : yAxisMap.values()) {
+    for (Axis_Y<ST, S> ya : yAxisMap.values()) {
       if (ya.getMin() == Double.POSITIVE_INFINITY || ya.getMax() == Double.POSITIVE_INFINITY) {
         throw new IllegalArgumentException(
             "Series data (accounting for error bars too) cannot be equal to Double.POSITIVE_INFINITY!!!");
@@ -361,12 +361,12 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     double xMin = chart.getStyler().isXAxisLogarithmic() ? 0.1 : -1.0;
     double yMin = chart.getStyler().isYAxisLogarithmic() ? 0.1 : -1.0;
     xAxis.addMinMax(xMin, 1);
-    for (Axis<ST, S> ya : yAxisMap.values()) {
+    for (Axis_Y<ST, S> ya : yAxisMap.values()) {
       ya.addMinMax(yMin, 1);
     }
   }
 
-  Axis<ST, S> getYAxis(int yIndex) {
+  Axis_Y<ST, S> getYAxis(int yIndex) {
 
     return yAxisMap.get(yIndex);
   }
@@ -399,7 +399,7 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     xAxis.setMax(overrideXAxisMaxValue);
   }
 
-  private void overrideMinMaxForYAxis(Axis yAxis) {
+  private void overrideMinMaxForYAxis(Axis_Y<ST, S> yAxis) {
 
     double overrideYAxisMinValue = yAxis.getMin();
     double overrideYAxisMaxValue = yAxis.getMax();
@@ -505,12 +505,12 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
 
   // Getters & Setters /////////////////////////////////////////////////
 
-  public Axis<ST, S> getXAxis() {
+  public Axis_X<ST, S> getXAxis() {
 
     return xAxis;
   }
 
-  Axis<ST, S> getYAxis() {
+  Axis_Y<ST, S> getYAxis() {
 
     return yAxis;
   }
@@ -531,12 +531,12 @@ public class AxisPair<ST extends AxesChartStyler, S extends AxesChartSeries> imp
     return rightYAxisBounds;
   }
 
-  Axis<ST, S> getLeftMainYAxis() {
+  Axis_Y<ST, S> getLeftMainYAxis() {
 
     return leftMainYAxis;
   }
 
-  Axis<ST, S> getRightMainYAxis() {
+  Axis_Y<ST, S> getRightMainYAxis() {
 
     return rightMainYAxis;
   }
