+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisTickCalculator_Callback.java
@@ -3,7 +3,7 @@ package org.knowm.xchart.internal.chartpart;
 import java.util.List;
 import java.util.function.Function;
 
-import org.knowm.xchart.internal.chartpart.Axis.Direction;
+import org.knowm.xchart.internal.chartpart.Axis_.Direction;
 import org.knowm.xchart.style.AxesChartStyler;
 
 /**
