+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/PlotContent_XY.java
@@ -61,7 +61,7 @@ public class PlotContent_XY<ST extends XYStyler, S extends XYSeries> extends Plo
       if (!series.isEnabled()) {
         continue;
       }
-      Axis yAxis = chart.getYAxis(series.getYAxisGroup());
+      Axis_Y yAxis = chart.getYAxis(series.getYAxisGroup());
       double yMin = yAxis.getMin();
       double yMax = yAxis.getMax();
       if (xyStyler.isYAxisLogarithmic()) {
